// SDKs/twilio.com/twilio_numbers_v2/1.37.3/typescript/src/sdk/models/shared/numbersv2regulatorycompliancebundlebundlecopy.ts
import { SpeakeasyMetadata, SpeakeasyBase } from "../../../internal/utils";
import { BundleCopyEnumStatusEnum } from "./bundlecopyenumstatusenum";



export class NumbersV2RegulatoryComplianceBundleBundleCopy extends SpeakeasyBase {
  @SpeakeasyMetadata({ data: "json, name=account_sid" })
  accountSid?: string;

  @SpeakeasyMetadata({ data: "json, name=date_created" })
  dateCreated?: Date;

  @SpeakeasyMetadata({ data: "json, name=date_updated" })
  dateUpdated?: Date;

  @SpeakeasyMetadata({ data: "json, name=email" })
  email?: string;

  @SpeakeasyMetadata({ data: "json, name=friendly_name" })
  friendlyName?: string;

  @SpeakeasyMetadata({ data: "json, name=regulation_sid" })
  regulationSid?: string;

  @SpeakeasyMetadata({ data: "json, name=sid" })
  sid?: string;

  @SpeakeasyMetadata({ data: "json, name=status" })
  status?: BundleCopyEnumStatusEnum;

  @SpeakeasyMetadata({ data: "json, name=status_callback" })
  statusCallback?: string;

  @SpeakeasyMetadata({ data: "json, name=valid_until" })
  validUntil?: Date;
}